requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
# langchain_community==0.0.29
langchain-openai==0.1.7
openai==1.27.0
python-dotenv==1.0.0
requests==2.31.0
unstructured==0.14.6
